+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Always show a popup as long as the active tab has a URL.
+- Detect dynamic changes to the canonical URL of a tab and update the icon face.
 
 ## 0.9.7